// src/services/api/generated/model/tenant.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * API
 * API docs
 * OpenAPI spec version: 1.0
 */
import type { TenantContactEmail } from './tenantContactEmail';
import type { TenantContactPhone } from './tenantContactPhone';
import type { TenantSettings } from './tenantSettings';

export interface Tenant {
  id: string;
  name: string;
  slug: string;
  contactEmail?: TenantContactEmail;
  contactPhone?: TenantContactPhone;
  isActive: boolean;
  settings?: TenantSettings;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string;
}