// src/services/api/generated/model/branch.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * API
 * API docs
 * OpenAPI spec version: 1.0
 */
import type { BranchAddress } from './branchAddress';
import type { BranchCity } from './branchCity';
import type { BranchCountry } from './branchCountry';
import type { BranchEmail } from './branchEmail';
import type { BranchPhone } from './branchPhone';
import type { BranchState } from './branchState';

export interface Branch {
  id: string;
  tenantId: string;
  name: string;
  code: string;
  address?: BranchAddress;
  city?: BranchCity;
  state?: BranchState;
  country?: BranchCountry;
  phone?: BranchPhone;
  email?: BranchEmail;
  isActive: boolean;
  isHeadquarters: boolean;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string;
}